Enable gosec G115: integer overflow conversion checks (#9302)
* Enable gosec G115: integer overflow conversion checks

  Remove G115 from the gosec excludes in .golangci.yml and fix all
  flagged conversion sites.

  True refactors:
  - version.From() now takes uint64 params directly, eliminating the
    int→uint64 conversions at the declaration site
  - verifier.ValidSignature and writeInt gain bounds checks that make
    the subsequent int→uint32 conversions provably safe
  - log.determineLogLevel clamps verbosity with max(*v*-1, math.MinInt8)
    to prevent a real int→int8 overflow for high verbosity values

  Remaining sites (node/pod/port counts, small fixed-range values) are
  annotated //nolint:gosec // G115: <reason> where static analysis
  cannot prove the range but overflow is not realistically possible.

* Use int32 parameter in WithNodeCount builder methods

  Aligns the parameter type with the underlying Spec.Count field type,
  removing the need for int→int32 conversions and their G115 annotations.

* Remove G115 exclusion from crypto.go

+++ test/e2e/test/enterprisesearch/builder.go
@@ -139,8 +139,8 @@ func (b Builder) WithVersion(version string) Builder {
 	return b
 }
 
-func (b Builder) WithNodeCount(count int) Builder {
-	b.EnterpriseSearch.Spec.Count = int32(count)
+func (b Builder) WithNodeCount(count int32) Builder {
+	b.EnterpriseSearch.Spec.Count = count
 	return b
 }
